requirements.txt
google-cloud-vision==0.39.0